Split matchmaker.algorithm into stable_marriage and bucket_teams
